- change column viewport form int to float - save viewport data to DB - load edges, nodes and flowChat from server side - open graph by flowchat

+++ database/migrations/2023_05_22_205719_create_flow_chats_table.php
@@ -16,9 +16,9 @@ return new class extends Migration
             $table->string('name');
             $table->text('description')->nullable();
             $table->boolean('active')->default(false);
-            $table->integer('viewport_x')->default(0);
-            $table->integer('viewport_y')->default(0);
-            $table->integer('viewport_zoom')->default(1);
+            $table->float('viewport_x')->default(0);
+            $table->float('viewport_y')->default(0);
+            $table->float('viewport_zoom')->default(1);
             $table->timestamps();
         });
     }
